1_2/DJ-HW-5/requirements.txt
Django
djangorestframework
psycopg2-binary
os
sys